+++ src/main/java/com/atsistemas/hotelBookings/Entity/Hotel.java
@@ -9,30 +9,32 @@ public class Hotel {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private Long id;
+    private Integer id;
     @Column(name = "name")
     private String name;
     @Column(name = "category")
     private String category;
+
     @OneToMany(mappedBy = "hotel", cascade = CascadeType.ALL, fetch = FetchType.LAZY)
     private List<Availability> availabilities;
+
 
     @OneToMany(mappedBy = "hotel", cascade = CascadeType.ALL, fetch = FetchType.LAZY)
     private List<Booking> bookings;
     public Hotel() {
     }
 
-    public Hotel(Long id, String name, String category) {
+    public Hotel(Integer id, String name, String category) {
         this.id = id;
         this.name = name;
         this.category = category;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
